bugfix: Fix uptime calculation in FreeRTOS port
Resolve issue in FreeRTOS port where quantization issues could occur with timer
tick values that are not even divisors of 1000.

For example, if a tick rate is 300Hz, due to the integer division to
ms_per_tick would be calculated to be 3ms (vs 3.33..3ms) which would cause
the later multiplication of ticks * ms_per_tick to result in an invalid ms
update.

Signed-off-by: Paul Buckley <paul@hubble.com>