crypto: Simplify aes ctr signature
The output data must be the size of the input data, which means that is
up to 13 bytes (which is less than AES block size). Change the function
signature to a pointer instead of an array of aes block size. This allow
using smaller buffer whithout need to cast it.

Signed-off-by: Flavio Ceolin <flavio@hubblenetwork.com>