ports: adjust files to support Arduino
In order to ease importing of the SDK to an Arduino library, make
some changes to the SDK primarily caused by Arduino:

  1. Compiling every compilation unit in the src/ dir of the lib
     (and not allowing sources/headers outside that)
  2. Not allowing build time defines to be passed to libraries

This necessitates so guards against compilation under some situations
and having a default set of configuration values. The SDK will still
need to be imported into the Arduino library repository but these
changes enable us to do this without changing the underlying files.

Signed-off-by: Paul Buckley <paul@hubble.com>